+++ package.json
@@ -28,7 +28,8 @@
     "react-native-toast-message": "^2.1.6",
     "react-native-web": "~0.18.7",
     "react-recompose": "^0.33.0",
-    "tailwind-rn": "^4.2.0"
+    "tailwind-rn": "^4.2.0",
+    "typescript": "^5.0.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
