+++ screens/listView.js
@@ -1,21 +1,17 @@
+import { useState, useEffect } from "react";
 import {
   Text,
   View,
-  Image,
   FlatList,
   useWindowDimensions,
   TouchableOpacity,
 } from "react-native";
 import { useTailwind } from "tailwind-rn";
-import { faker } from "@faker-js/faker";
-import { useContext, useEffect, useState } from "react";
 import { useNavigation } from "@react-navigation/native";
 import { ListViewCard } from "../components/cards";
 import PageContainer from "../components/pageContainer";
 import { getBreakPoint } from "../utils/screen";
 import { search } from "../api/search";
-import { addressDetailsContext } from "../contexts/AddressContext";
-import SearchBarComponent from "../components/searchBar";
 
 const orderByDescending = (a, b) => b - a;
 const orderByAscending = (a, b) => a - b;
@@ -52,11 +48,10 @@ const ListView = ({ route }) => {
       orderBy: orderByDescending,
     },
   ];
-  const [foodTypeScreen, showFoodTypeScreen] = useState(false);
+
   const numColumns = { sm: 2, md: 3, lg: 4, xl: 5 };
   const window = useWindowDimensions();
   const [searchResults, setSearchResults] = useState([]);
-  const address = useContext(addressDetailsContext);
   const [showFilters, setShowFilters] = useState(false);
 
   const filterResults = (results) =>
@@ -76,42 +71,6 @@ const ListView = ({ route }) => {
 
   return (
     <PageContainer style={tailwind("m-2")}>
-      <View>
-        <View style={tailwind("flex flex-row justify-between")}>
-          <View>
-            <Text style={tailwind("text-3xl font-bold")}>
-              Hello {faker.name.firstName()} 👋
-            </Text>
-            <Text style={tailwind("text-xl font-light")}>
-              {"It's lunch time!"}
-            </Text>
-          </View>
-          <Image
-            style={tailwind("w-5 h-5")}
-            resizeMode="contain"
-            source={require("../assets/icons/black/search.png")}
-          />
-        </View>
-        <View style={tailwind("my-3")}>
-          <Text style={tailwind("text-2xl font-bold")}>Nearby food</Text>
-          <View style={tailwind("flex flex-row items-center")}>
-            <Image
-              style={tailwind("w-4 h-4")}
-              resizeMode="contain"
-              source={require("../assets/icons/black/location.png")}
-            />
-            <Text style={tailwind("font-light ml-2")}>
-              {address[0].address.address1}
-            </Text>
-          </View>
-        </View>
-        <View style={tailwind("flex flex-row items-center")}>
-          <SearchBarComponent
-            isFoodTypesOpen={foodTypeScreen}
-            openFoodTypes={showFoodTypeScreen}
-          />
-        </View>
-      </View>
       <TouchableOpacity
         style={tailwind(
           `${
