+++ tailwind.css
@@ -14,10 +14,6 @@
   position: absolute
 }
 
-.relative {
-  position: relative
-}
-
 .top-0 {
   top: 0px
 }
@@ -43,14 +39,19 @@
   margin-right: 0.5rem
 }
 
+.my-4 {
+  margin-top: 1rem;
+  margin-bottom: 1rem
+}
+
 .my-3 {
   margin-top: 0.75rem;
   margin-bottom: 0.75rem
 }
 
-.my-4 {
-  margin-top: 1rem;
-  margin-bottom: 1rem
+.my-2 {
+  margin-top: 0.5rem;
+  margin-bottom: 0.5rem
 }
 
 .mx-4 {
@@ -70,6 +71,10 @@
   margin-right: 0.25rem
 }
 
+.mt-\[100px\] {
+  margin-top: 100px
+}
+
 .mt-2 {
   margin-top: 0.5rem
 }
@@ -82,16 +87,12 @@
   margin-top: 1rem
 }
 
-.mr-4 {
-  margin-right: 1rem
-}
-
 .ml-2 {
   margin-left: 0.5rem
 }
 
-.mt-\[100px\] {
-  margin-top: 100px
+.mr-4 {
+  margin-right: 1rem
 }
 
 .flex {
@@ -114,6 +115,10 @@
   height: 250px
 }
 
+.h-\[100px\] {
+  height: 100px
+}
+
 .h-10 {
   height: 2.5rem
 }
@@ -122,8 +127,8 @@
   height: 6rem
 }
 
-.h-9 {
-  height: 2.25rem
+.h-\[200px\] {
+  height: 200px
 }
 
 .h-4 {
@@ -134,24 +139,32 @@
   height: 1.25rem
 }
 
-.h-\[200px\] {
-  height: 200px
+.h-9 {
+  height: 2.25rem
 }
 
 .h-full {
   height: 100%
 }
 
-.h-\[26px\] {
-  height: 26px
+.h-20 {
+  height: 5rem
 }
 
-.h-\[50px\] {
-  height: 50px
+.h-11 {
+  height: 2.75rem
 }
 
-.h-\[100px\] {
-  height: 100px
+.h-12 {
+  height: 3rem
+}
+
+.h-14 {
+  height: 3.5rem
+}
+
+.h-36 {
+  height: 9rem
 }
 
 .min-h-\[320px\] {
@@ -182,8 +195,8 @@
   width: 6rem
 }
 
-.w-9 {
-  width: 2.25rem
+.w-10 {
+  width: 2.5rem
 }
 
 .w-4 {
@@ -194,8 +207,28 @@
   width: 1.25rem
 }
 
-.w-10 {
-  width: 2.5rem
+.w-9 {
+  width: 2.25rem
+}
+
+.w-20 {
+  width: 5rem
+}
+
+.w-28 {
+  width: 7rem
+}
+
+.w-32 {
+  width: 8rem
+}
+
+.w-36 {
+  width: 9rem
+}
+
+.w-14 {
+  width: 3.5rem
 }
 
 .flex-1 {
@@ -270,6 +303,10 @@
   background-color: rgb(243 244 246 / var(--tw-bg-opacity))
 }
 
+.bg-white\/90 {
+  background-color: rgb(255 255 255 / 0.9)
+}
+
 .bg-gray-200 {
   --tw-bg-opacity: 1;
   background-color: rgb(229 231 235 / var(--tw-bg-opacity))
@@ -287,10 +324,6 @@
 
 .bg-black\/40 {
   background-color: rgb(0 0 0 / 0.4)
-}
-
-.bg-white\/90 {
-  background-color: rgb(255 255 255 / 0.9)
 }
 
 .bg-none {
@@ -317,10 +350,6 @@
   padding: 0.625rem
 }
 
-.pt-20 {
-  padding-top: 5rem
-}
-
 .pb-20 {
   padding-bottom: 5rem
 }
@@ -329,12 +358,8 @@
   padding-top: 0.5rem
 }
 
-.pt-\[100px\] {
-  padding-top: 100px
-}
-
-.pt-\[20px\] {
-  padding-top: 20px
+.pb-4 {
+  padding-bottom: 1rem
 }
 
 .text-center {
@@ -375,12 +400,12 @@
   font-weight: 700
 }
 
-.font-light {
-  font-weight: 300
-}
-
 .font-medium {
   font-weight: 500
+}
+
+.font-light {
+  font-weight: 300
 }
 
 .font-extrabold {
