+++ tailwind.json
@@ -19,11 +19,6 @@
 			"position": "absolute"
 		}
 	},
-	"relative": {
-		"style": {
-			"position": "relative"
-		}
-	},
 	"top-0": {
 		"style": {
 			"top": 0
@@ -64,16 +59,22 @@
 			"marginRight": 8
 		}
 	},
+	"my-4": {
+		"style": {
+			"marginTop": 16,
+			"marginBottom": 16
+		}
+	},
 	"my-3": {
 		"style": {
 			"marginTop": 12,
 			"marginBottom": 12
 		}
 	},
-	"my-4": {
+	"my-2": {
 		"style": {
-			"marginTop": 16,
-			"marginBottom": 16
+			"marginTop": 8,
+			"marginBottom": 8
 		}
 	},
 	"mx-4": {
@@ -97,6 +98,11 @@
 			"marginRight": 4
 		}
 	},
+	"mt-[100px]": {
+		"style": {
+			"marginTop": 100
+		}
+	},
 	"mt-2": {
 		"style": {
 			"marginTop": 8
@@ -112,19 +118,14 @@
 			"marginTop": 16
 		}
 	},
-	"mr-4": {
-		"style": {
-			"marginRight": 16
-		}
-	},
 	"ml-2": {
 		"style": {
 			"marginLeft": 8
 		}
 	},
-	"mt-[100px]": {
+	"mr-4": {
 		"style": {
-			"marginTop": 100
+			"marginRight": 16
 		}
 	},
 	"flex": {
@@ -152,6 +153,11 @@
 			"height": 250
 		}
 	},
+	"h-[100px]": {
+		"style": {
+			"height": 100
+		}
+	},
 	"h-10": {
 		"style": {
 			"height": 40
@@ -162,9 +168,9 @@
 			"height": 96
 		}
 	},
-	"h-9": {
+	"h-[200px]": {
 		"style": {
-			"height": 36
+			"height": 200
 		}
 	},
 	"h-4": {
@@ -177,9 +183,9 @@
 			"height": 20
 		}
 	},
-	"h-[200px]": {
+	"h-9": {
 		"style": {
-			"height": 200
+			"height": 36
 		}
 	},
 	"h-full": {
@@ -187,19 +193,29 @@
 			"height": "100%"
 		}
 	},
-	"h-[26px]": {
+	"h-20": {
 		"style": {
-			"height": 26
+			"height": 80
 		}
 	},
-	"h-[50px]": {
+	"h-11": {
 		"style": {
-			"height": 50
+			"height": 44
 		}
 	},
-	"h-[100px]": {
+	"h-12": {
 		"style": {
-			"height": 100
+			"height": 48
+		}
+	},
+	"h-14": {
+		"style": {
+			"height": 56
+		}
+	},
+	"h-36": {
+		"style": {
+			"height": 144
 		}
 	},
 	"min-h-[320px]": {
@@ -237,9 +253,9 @@
 			"width": 96
 		}
 	},
-	"w-9": {
+	"w-10": {
 		"style": {
-			"width": 36
+			"width": 40
 		}
 	},
 	"w-4": {
@@ -252,9 +268,34 @@
 			"width": 20
 		}
 	},
-	"w-10": {
+	"w-9": {
 		"style": {
-			"width": 40
+			"width": 36
+		}
+	},
+	"w-20": {
+		"style": {
+			"width": 80
+		}
+	},
+	"w-28": {
+		"style": {
+			"width": 112
+		}
+	},
+	"w-32": {
+		"style": {
+			"width": 128
+		}
+	},
+	"w-36": {
+		"style": {
+			"width": 144
+		}
+	},
+	"w-14": {
+		"style": {
+			"width": 56
 		}
 	},
 	"flex-1": {
@@ -355,6 +396,11 @@
 			"backgroundColor": "rgb(243 244 246 / var(--tw-bg-opacity))"
 		}
 	},
+	"bg-white/90": {
+		"style": {
+			"backgroundColor": "rgb(255 255 255 / 0.9)"
+		}
+	},
 	"bg-gray-200": {
 		"style": {
 			"--tw-bg-opacity": 1,
@@ -376,11 +422,6 @@
 	"bg-black/40": {
 		"style": {
 			"backgroundColor": "rgb(0 0 0 / 0.4)"
-		}
-	},
-	"bg-white/90": {
-		"style": {
-			"backgroundColor": "rgb(255 255 255 / 0.9)"
 		}
 	},
 	"bg-none": {
@@ -425,11 +466,6 @@
 			"paddingLeft": 10
 		}
 	},
-	"pt-20": {
-		"style": {
-			"paddingTop": 80
-		}
-	},
 	"pb-20": {
 		"style": {
 			"paddingBottom": 80
@@ -440,14 +476,9 @@
 			"paddingTop": 8
 		}
 	},
-	"pt-[100px]": {
+	"pb-4": {
 		"style": {
-			"paddingTop": 100
-		}
-	},
-	"pt-[20px]": {
-		"style": {
-			"paddingTop": 20
+			"paddingBottom": 16
 		}
 	},
 	"text-center": {
@@ -496,14 +527,14 @@
 			"fontWeight": "700"
 		}
 	},
-	"font-light": {
-		"style": {
-			"fontWeight": "300"
-		}
-	},
 	"font-medium": {
 		"style": {
 			"fontWeight": "500"
+		}
+	},
+	"font-light": {
+		"style": {
+			"fontWeight": "300"
 		}
 	},
 	"font-extrabold": {
